+++ include/caffe/common_layers.hpp
@@ -735,7 +735,7 @@ class BNLayer : public Layer<Dtype> {
 
 
 #if defined(USE_CUDNN) 
-#if CUDNN_VERSION_MIN(4, 0, 0)
+#if CUDNN_VERSION_MIN(5, 0, 0)
 /**
  * @brief cuDNN implementation of BNLayer.
  *        Fallback to BNLayer for CPU mode.
@@ -767,8 +767,6 @@ class CuDNNBNLayer : public BNLayer<Dtype> {
   cudnnTensorDescriptor_t top_desc_;
   cudnnTensorDescriptor_t bn_param_desc_;
 
-  Blob<Dtype> scale_buf_;
-  Blob<Dtype> bias_buf_;
   Blob<Dtype> save_mean_;
   Blob<Dtype> save_inv_variance_;
 };