+++ src/caffe/layer_factory.cpp
@@ -168,35 +168,23 @@ shared_ptr<Layer<Dtype> > GetBNLayer(const LayerParameter& param) {
   if (engine == BNParameter_Engine_DEFAULT) {
     engine = BNParameter_Engine_CAFFE;
 #if defined(USE_CUDNN)
-#if CUDNN_VERSION_MIN(4, 0, 0)
-    // TODO : Currently we use caffe as the default engine, due to the performance issues with NV's implementation.
-    // Will switch back when this get fixed.
-
-    // engine = BNParameter_Engine_CUDNN;
+#if CUDNN_VERSION_MIN(5, 0, 0)
+     engine = BNParameter_Engine_CUDNN;
 #endif
 #endif
   }
-#if defined(USE_CUDNN)
-#if CUDNN_VERSION_MIN(4, 0, 0)
-  if (engine == BNParameter_Engine_CUDNN && param.bn_param().frozen()) {
-    LOG(WARNING) << "Layer " << param.name() << " switches back to CAFFE engine"
-                 << " as CUDNN engine doesn't support frozen.";
-    engine = BNParameter_Engine_CAFFE;
-  }
-#endif
-#endif
   if (engine == BNParameter_Engine_CAFFE) {
     LOG(INFO) << "Layer " << param.name() << " is using CAFFE engine.";
     return shared_ptr<Layer<Dtype> >(new BNLayer<Dtype>(param));
 #if defined(USE_CUDNN)
-#if CUDNN_VERSION_MIN(4, 0, 0)
+#if CUDNN_VERSION_MIN(5, 0, 0)
   } else if (engine == BNParameter_Engine_CUDNN) {
     LOG(INFO) << "Layer " << param.name() << " is using CUDNN engine.";
     return shared_ptr<Layer<Dtype> >(new CuDNNBNLayer<Dtype>(param));
 #endif
 #endif
   } else {
-    LOG(FATAL) << "Layer " << param.name() << " has unknown engine.";
+    LOG(FATAL) << "Layer " << param.name() << " calls cuDNN engine, but cuDNN version higher than 5.0 is not found";
   }
 }